+++ js/util.js
@@ -1,0 +1,2 @@
+const getRandomNumber = (min,max) => Math.floor(Math.random() * (Math.max(max,min) - Math.min(max,min) + 1)) + Math.min(max,min);
+export {getRandomNumber};
